Create Pascal triangle using recursion and memoization
In this code we use a HashMap called memo to store previously calculated values of Pascal Triangle. The calculatePascalValue method checks if the value for the given row and column is already present in the memo. If it is, it returns the stored value. otherwise it calculates the value using recursion by adding the values from the row above (row - 1). The calculated value is then stored in the memo for future use.

The main method prints Pascal Triangle by iterating through each row and column calling the calculatePascalValue method and displays the calculated value.